# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hawk-Eye Control Plane",
    "description": "Insider & privileged-user fraud detection — FastAPI control plane (ALERT-ONLY).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_rules_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleSummary"
                  },
                  "title": "Response List Rules Api V1 Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Propose Rule Change",
        "description": "Propose a rule/threshold change. Pending until a second approver signs off (four-eyes).",
        "operationId": "propose_rule_change_api_v1_rules_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleChangeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleChangeProposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/{code}": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Update Rule Change",
        "description": "PUT = propose an *update* to an existing rule (Part 24.2 `PUT /rules`). Like POST, it is a\nfour-eyes proposal: it takes effect only after a second approver signs off, is versioned, and\nis audited. The rule code is taken from the path.",
        "operationId": "update_rule_change_api_v1_rules__code__put",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleChangeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleChangeProposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/{change_id}/approve": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Approve Rule Change",
        "operationId": "approve_rule_change_api_v1_rules__change_id__approve_post",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleApproval"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleApprovalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/fmr": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report Fmr",
        "operationId": "report_fmr_api_v1_reports_fmr_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FmrReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/crilc": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report Crilc",
        "operationId": "report_crilc_api_v1_reports_crilc_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrilcReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrilcLineItem": {
        "properties": {
          "crilc_id": {
            "type": "string",
            "title": "Crilc Id"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "exposure_inr": {
            "type": "integer",
            "title": "Exposure Inr"
          },
          "crosses_3cr": {
            "type": "boolean",
            "title": "Crosses 3Cr"
          },
          "reporting_due_ts": {
            "type": "string",
            "title": "Reporting Due Ts",
            "description": "7-day reporting trigger deadline"
          },
          "classification_due_ts": {
            "type": "string",
            "title": "Classification Due Ts",
            "description": "180-day classification window end"
          },
          "rfa_tagged": {
            "type": "boolean",
            "title": "Rfa Tagged",
            "default": false
          },
          "sma_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sma Class",
            "description": "Special Mention Account class, if applicable"
          }
        },
        "type": "object",
        "required": [
          "crilc_id",
          "entity_id",
          "exposure_inr",
          "crosses_3cr",
          "reporting_due_ts",
          "classification_due_ts"
        ],
        "title": "CrilcLineItem",
        "description": "One CRILC line — ₹3-crore / 7-day trigger, 180-day classification window."
      },
      "CrilcReport": {
        "properties": {
          "generated_ts": {
            "type": "string",
            "title": "Generated Ts"
          },
          "submission_enabled": {
            "type": "boolean",
            "title": "Submission Enabled",
            "default": false
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CrilcLineItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_exposure_inr": {
            "type": "integer",
            "title": "Total Exposure Inr",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "generated_ts"
        ],
        "title": "CrilcReport"
      },
...
      "FmrLineItem": {
        "properties": {
          "fmr_id": {
            "type": "string",
            "title": "Fmr Id"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "alert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "RBI FMR fraud category (synthetic mapping)"
          },
          "amount_inr": {
            "type": "integer",
            "title": "Amount Inr"
          },
          "detected_ts": {
            "type": "string",
            "title": "Detected Ts"
          },
          "rfa_tagged": {
            "type": "boolean",
            "title": "Rfa Tagged",
            "default": false
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "reported"
          }
        },
        "type": "object",
        "required": [
          "fmr_id",
          "entity_id",
          "category",
          "amount_inr",
          "detected_ts"
        ],
        "title": "FmrLineItem",
        "description": "One Fraud Monitoring Return line (synthetic)."
      },
      "FmrReport": {
        "properties": {
          "generated_ts": {
            "type": "string",
            "title": "Generated Ts"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "submission_enabled": {
            "type": "boolean",
            "title": "Submission Enabled",
            "description": "SCAFFOLD: live RBI channel absent locally",
            "default": false
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FmrLineItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_amount_inr": {
            "type": "integer",
            "title": "Total Amount Inr",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "generated_ts",
          "period"
        ],
        "title": "FmrReport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleApproval": {
        "properties": {
          "change_id": {
            "type": "string",
            "title": "Change Id"
          },
          "approve": {
            "type": "boolean",
            "title": "Approve",
            "default": true
          },
          "approver_notes": {
            "type": "string",
            "title": "Approver Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "change_id"
        ],
        "title": "RuleApproval"
      },
      "RuleApprovalResult": {
        "properties": {
          "change_id": {
            "type": "string",
            "title": "Change Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "new_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Version"
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          }
        },
        "type": "object",
        "required": [
          "change_id",
          "code",
          "status",
          "audit_id"
        ],
        "title": "RuleApprovalResult"
      },
      "RuleChangeProposal": {
        "properties": {
          "change_id": {
            "type": "string",
            "title": "Change Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "pending_approval | approved | rejected",
            "default": "pending_approval"
          },
          "diff": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diff"
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          }
        },
        "type": "object",
        "required": [
          "change_id",
          "code",
          "proposed_by",
          "audit_id"
        ],
        "title": "RuleChangeProposal"
      },
      "RuleChangeRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Severity"
              },
              {
                "type": "null"
              }
            ]
          },
          "hard_hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hard Hit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "change_reason": {
            "type": "string",
            "title": "Change Reason",
            "description": "Why the change is needed (audited)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "change_reason"
        ],
        "title": "RuleChangeRequest",
        "description": "A *proposed* change — does not take effect until a second approver signs off (four-eyes)."
      },
...
      "RuleSummary": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "NEW_BENEFICIARY_THEN_HIGHVALUE"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "hard_hit": {
            "type": "boolean",
            "title": "Hard Hit",
            "description": "Hard-hit rules feed the L1 short-circuit",
            "default": false
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "version",
          "enabled",
          "severity"
        ],
        "title": "RuleSummary"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}